sm8150-common: Use dynamic partitions
The main purpose of this is to have an odm partition and make use of the
(currently unused) oem partition. Using odm allows us to run some scripts
that from QPR2 onwards were unable to run (most notably, the msim script).

Do note that due to BOARD_SHIPPING_API_LEVEL := 30 this is not
straightforward to retrofit and requires a patched update_engine to remove
a (bothersome) check. Kernel patches are also required due to broken
system-as-root-enforcing bootloader arguments.

Signed-off-by: voidanix <voidanix@keyedlimepie.org>